requirements.txt

# AuraDocs Full-Stack Dependencies
# To install Backend (Python): cd backend && pip install -r ../requirements.txt 
# To install Frontend (Node.js): cd frontend && npm install ... (or use package.json)

# ==========================================
# 🐍 BACKEND DEPENDENCIES (Python / Django)
# ==========================================

# Web Framework & API
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0

# Database
PyMySQL==1.1.2

# Web Scraping & Metadata Extraction
beautifulsoup4==4.14.3
selenium==4.43.0
webdriver-manager==4.0.2

# Vector Database (RAG) & Embeddings
chromadb==1.5.8
sentence-transformers==5.4.1

# AI & LLM Inference
openai==2.32.0

# Utilities
python-dotenv==1.2.2
requests==2.33.1


# ==========================================
# ⚛️ FRONTEND DEPENDENCIES (Node.js / Next.js)
# ==========================================
# (These represent matching package.json versions)

# Core Framework
next==16.2.4
react==19.2.4
react-dom==19.2.4

# API & Networking
axios>=1.15.0

# Styling, Animations & UI
tailwindcss>=4.0.0
@tailwindcss/postcss>=4.0.0
framer-motion>=12.38.0
lucide-react>=1.8.0

# TypeScript & Developer Tools
typescript>=5.0.0
eslint>=9.0.0
eslint-config-next==16.2.4
@types/node>=20.0.0
@types/react>=19.0.0
@types/react-dom>=19.0.0
